Move PointAdder contents to Tree
