+++ Rust/src/tree/mod.rs
@@ -9,11 +9,8 @@ pub use cut::Cut;
 mod node;
 pub use node::{Internal, Leaf, Node};
 
-mod point_adder;
-pub use point_adder::{AddResult, PointAdder};
-
 mod point_deleter;
 pub use point_deleter::{DeleteResult, PointDeleter};
 
 mod tree;
-pub use tree::{NodeTraverser, Tree};
+pub use tree::{AddResult, NodeTraverser, Tree};
